refactor Gas_container's constructor
Gas_container() now creates and stores a ci::Rectf object which is used to draw every loop. Moving this to a member variable allows other objects within the container to access the boundary rect and it additionally saves from having to reconstructor the rect each time the draw() function is called.