debug inter-particle collisions
Inter-particle collision bugs such as wall phasing and particle sticking is now accounted for. The bugs were a little more insidious than previously thought. Rather than being a problem with the implementation of the particle directionality equation or the inelastic collision equation, it was a problem with updating particle state. All particles needed to be moved to their final locations and then collisions could be handled. Previously, all particles were moved individually and collisions calculated during the move of that single particle. On top of this, there was another bug of updating the state of one colliding particle and then using that newly updated state of that particle to update the state of the colliding particle's collision partner. The state of the two colliding particles (position & velocity) needed to be saved and then used to alter each individual state.