convert particle references to pointers
Particle objects are now now being tracked with pointers and the pointers are stored in vectors inside gas_container.cc. This allows for simultaneous storage of a single particle object in multiple locations. Now all particles can be stored in a a general vector of particles while also being stored in containers that are specific to their type/color. Changes made to all the particles during handling of collisions is reflected in the other storage containers. Storing single objects in multiple places allows for easier separation of objects when feeding into gas_histogram.cc's Update() function.