Implement x-axis sharing amongst histograms.
In order for each histogram to have the same shared x axis, they all need to agree on the maximum velocity for creating the appropriate bin sizes. Prior to this a static variable was used to share this information amongst classes. This resulted in a bug where the first histogram would calculate a max speed based on the particles presented to it and create bins, then the next histogram would continue to update the max speed variable and create bins, etc., etc. This however does not take into account that histogram one has already set its bin width (and x-axis) which will surely be updated later when the other histograms are updated. This is a slightly insidious problem. In order to preserve encapsulation, each histogram is given only the necessary particles to generate it's graph and no other extraneous information. In order to get the true max velocity of all the particles however, it would need access to all the particles (not just the ones it intends to put into bins). This is worked around by calculating a the max speed outside of the histogram. This functionality now takes place inside the gas_container class and I feel that this is a more natural place for such a calculation. Now gas_histogram's Update() function not only takes in the vector of particles which it is to place into bins but also the max velocity of that turn so that it can create the appropriate, shared x-axis.