fix bug associated with particle phasing through container boundary
Prior to this fix, collisions were handled based on their directionality (i.e. if the two objects were heading towards or away from each other). This is useful as it keeps particles from immediately reversing directions upon a collision and sticking to the walls of the container or other particles. This equation however, becomes a problem when somehow a particle has gained enough velocity to place its center of mass past the wall boundary. In this situation, the equation for directionality keeps it from switching directions and the particles floats off into space. When this directionality equation is removed, particles no longer float off into space once their center of mass is past the boundary. Instead, they enter into a state of limbo where their exit velocity was enough to push them well beyond the boundary but their return velocity after switch directions isn't enough to return them to the box (they are essentially stuck out in space and don't ask me how this is possible since collisions are elastic and handled in a step-wise fashion). The fix to this (or maybe it is more a band-aid since it appears my simulation is defying some laws of elastic collisions) is to apply the directionality equation but only in certain circumstances. 1) When inside the box, apply the equation as intended 2) When the center of mass is outside the box, apply it in reverse (i.e. only reverse direction if the container boundary and particle are heading away!). This seems to be a good fix and now particles will not phase through completely.